# БР1.2/Москалец Данила/homeworks/hw2/openapi.yaml
openapi: 3.0.0
info:
  title: Sangel API
  description: Маркетплейс охранных услуг
  version: 1.0.0
  contact:
    name: Москалец Данила Алексеевич
servers:
  - url: https://api.sangel.ru/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean

    # User
...
    Company:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        logo:
          type: string
        website:
          type: string
        avg_rating:
          type: number
          format: float
        total_reviews:
          type: integer
        created_at:
          type: string
          format: date-time
    CompanyDetail:
      allOf:
        - $ref: '#/components/schemas/Company'
        - type: object
          properties:
            owner:
              type: object
              properties:
                id:
                  type: integer
                first_name:
                  type: string
                last_name:
                  type: string
            services:
              type: array
              items:
                $ref: '#/components/schemas/Service'
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
    CreateCompanyRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        logo:
          type: string
        website:
          type: string

    # Category
    Category:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        is_published:
          type: boolean
...
    Service:
      type: object
      properties:
        id:
          type: integer
        company_id:
          type: integer
        company_title:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: decimal
        final_price:
          type: number
          format: decimal
        discount_percentage:
          type: integer
        is_published:
          type: boolean
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        avg_rating:
          type: number
          format: float
...
    Discount:
      type: object
      properties:
        id:
          type: integer
        service_id:
          type: integer
        percentage:
          type: integer
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
    CreateDiscountRequest:
      type: object
      required:
        - percentage
        - start_at
        - end_at
      properties:
        percentage:
          type: integer
          minimum: 1
          maximum: 100
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time

    # Request
...
    Review:
      type: object
      properties:
        id:
          type: integer
        service_id:
          type: integer
        service_name:
          type: string
        company_id:
          type: integer
        company_title:
          type: string
        user:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        created_at:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /companies:
    get:
      summary: Список компаний
      tags: [Companies]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: integer
        - name: price_min
          in: query
          schema:
            type: number
        - name: price_max
          in: query
          schema:
            type: number
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [price, rating]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Создать компанию
      tags: [Companies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        201:
          description: Компания создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Company'
        403:
          description: Пользователь уже владеет компанией
  /companies/{id}:
    get:
      summary: Детальная информация о компании
      tags: [Companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CompanyDetail'
        404:
          description: Компания не найдена

    put:
      summary: Обновить компанию
      tags: [Companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        200:
          description: Компания обновлена
        403:
          description: Доступ запрещен
        404:
          description: Компания не найдена

    delete:
      summary: Удалить компанию
      tags: [Companies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Компания удалена
        403:
          description: Доступ запрещен
        404:
          description: Компания не найдена


  # ==================== CATEGORIES ====================
...
  /services/{service_id}/discount:
    get:
      summary: Получить скидку услуги
      tags: [Discounts]
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Успешно
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Discount'
        404:
          description: Скидка не найдена

    post:
      summary: Создать скидку
      tags: [Discounts]
      parameters:
        - name: service_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiscountRequest'
      responses:
        201:
          description: Скидка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Discount'
        403:
          description: Доступ запрещен
        404:
          description: Услуга не найдена
        409:
          description: У услуги уже есть активная скидка
  /discounts/{id}:
    put:
      summary: Обновить скидку
      tags: [Discounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDiscountRequest'
      responses:
        200:
          description: Скидка обновлена
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Discount'
        403:
          description: Доступ запрещен
        404:
          description: Скидка не найдена

    delete:
      summary: Удалить скидку
      tags: [Discounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Скидка удалена
        403:
          description: Доступ запрещен
        404:
          description: Скидка не найдена

  # ==================== REQUESTS ====================
...